+++ conf.py
@@ -62,5 +62,4 @@ html_static_path = ['_static']
 html_css_files = ['custom.css']
 
 
-
 html_show_sourcelink = False