+++ tmart/AEC/__init__.py
@@ -33,5 +33,3 @@ from .run_regular import *
 from .run import *
 from .unzip import *
 from .write_atm_info import *
-
-