+++ tmart/AEC/identify_input.py
@@ -11,10 +11,6 @@
 # With contribution from Shun Bi
 
 def identify_input(file_path):
-    '''
-    Identify the type of the input: directory, file, or .SAFE.zip
-    
-    '''
     import os, sys
     import tmart