+++ tmart/__init__.py
@@ -9,9 +9,7 @@
 
 from .Surface import Surface, SpectralSurface
 from .Atmosphere import Atmosphere
-
 from .tmart import Tmart
-
 from .tm_calcref import calc_ref 
 from .tm_geometry import dirP_to_coord